pyproject.toml
[project]
name = "fastapi-2-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "orjson>=3.11.3",
    "pendulum>=3.1.0",
    "pre-commit>=4.3.0",
    "pydantic-extra-types[pendulum]>=2.10.5",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "ruff>=0.12.11",
    "scalar-fastapi>=1.3.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
